refactor: refactor CallbackMetric Protocol and move UnavailableDeploymentTotalMetric to paasng
N
