+++ apiserver/paasng/paasng/metrics/collector.py
@@ -16,14 +16,15 @@ limitations under the License.
 We undertake not to change the open source license (MIT license) applicable
 to the current version of the project delivered to anyone in the future.
 """
-from typing import Any, Dict, Type
+from typing import Dict, Iterator, Type
 
-from prometheus_client.core import GaugeMetricFamily
+from prometheus_client.core import Metric
 from typing_extensions import Protocol
 
 from paasng.metrics.basic_services.blob_store import BlobStoreAvailableMetric
 from paasng.metrics.basic_services.mysql import MySQLAvailableMetric
 from paasng.metrics.basic_services.redis import RedisAvailableMetric
+from paasng.metrics.workloads.deployment import UnavailableDeploymentTotalMetric
 
 
 class CallbackMetric(Protocol):
@@ -31,17 +32,19 @@ class CallbackMetric(Protocol):
     支持回调求值的 Metric
 
     :param name: 指标名
-    :param metric_type: 指标类型
     :param description: 指标描述
     """
 
     name: str
-    metric_type: str
     description: str
 
     @classmethod
-    def calc_value(cls) -> Any:
-        """获取 metric 值"""
+    def calc_metric(cls) -> Metric:
+        """获取 metric"""
+
+    @classmethod
+    def describe_metric(cls) -> Metric:
+        """描述 metric"""
 
 
 class CallbackMetricCollector:
@@ -54,11 +57,13 @@ class CallbackMetricCollector:
         self._metrics[metric.name] = metric
 
     def collect(self):
-        for metric in self._metrics.values():
-            if metric.metric_type == 'gauge':
-                yield GaugeMetricFamily(metric.name, metric.description, metric.calc_value())
-            else:
-                raise ValueError('CallbackMetricCollector only support add gauge metric at this time')
+        for m in self._metrics.values():
+            yield m.calc_metric()
+
+    def describe(self) -> Iterator[Metric]:
+        """describe the metrics that this collector provided"""
+        for m in self._metrics.values():
+            yield m.describe_metric()
 
 
 cb_metric_collector = CallbackMetricCollector()
@@ -67,3 +72,5 @@ cb_metric_collector = CallbackMetricCollector()
 cb_metric_collector.add(BlobStoreAvailableMetric)
 cb_metric_collector.add(MySQLAvailableMetric)
 cb_metric_collector.add(RedisAvailableMetric)
+# 添加原 workloads metric 指标
+cb_metric_collector.add(UnavailableDeploymentTotalMetric)
